Refactor to include generation code out of the nearley-unparse script into the library module so that user code can have access to it.
